[+] Some AHBot changes for buyout prices. [+] Added a Ulduar Boss to Scriptmgr. [+] Update and merge to TC2 rev 4521

+++ src/game/SkillDiscovery.cpp
@@ -26,6 +26,7 @@
 #include "Util.h"
 #include "SkillDiscovery.h"
 #include "SpellMgr.h"
+#include "Player.h"
 #include <map>
 
 struct SkillDiscoveryEntry
@@ -37,7 +38,7 @@ struct SkillDiscoveryEntry
     SkillDiscoveryEntry()
         : spellId(0), reqSkillValue(0), chance(0) {}
 
-    SkillDiscoveryEntry(uint16 _spellId, uint32 req_skill_val, float _chance)
+    SkillDiscoveryEntry(uint32 _spellId, uint32 req_skill_val, float _chance)
         : spellId(_spellId), reqSkillValue(req_skill_val), chance(_chance) {}
 };
